+++ compile/depack.js
@@ -55,10 +55,10 @@ function sa(a, b, c = !1) {
 };
 class va extends ka {
   constructor(a) {
-    const {binary:b = !1, rs:c = null, ...d} = a || {}, {S:e = B(!0), proxyError:f} = a || {}, g = (k, h) => e(h);
+    const {binary:b = !1, rs:c = null, ...d} = a || {}, {K:e = B(!0), proxyError:f} = a || {}, g = (k, h) => e(h);
     super(d);
     this.a = [];
-    this.M = new Promise((k, h) => {
+    this.F = new Promise((k, h) => {
       this.on("finish", () => {
         let l;
         b ? l = Buffer.concat(this.a) : l = this.a.join("");
@@ -83,11 +83,11 @@ class va extends ka {
     c();
   }
   get promise() {
-    return this.M;
+    return this.F;
   }
 }
 const C = async a => {
-  ({promise:a} = new va({rs:a, S:B(!0)}));
+  ({promise:a} = new va({rs:a, K:B(!0)}));
   return await a;
 };
 async function D(a) {
@@ -129,7 +129,7 @@ async function F(a, b, c) {
     a(...h);
   });
 }
-;const {basename:H, dirname:J, join:K, relative:M, resolve:xa, sep:ya} = path;
+;const {basename:G, dirname:J, join:K, relative:M, resolve:xa, sep:ya} = path;
 async function za(a) {
   const b = J(a);
   try {
@@ -251,7 +251,7 @@ const Q = async(a, b) => {
       }
     }
   }
-  return {path:a.startsWith(".") ? M("", b) : b, aa:d};
+  return {path:a.startsWith(".") ? M("", b) : b, U:d};
 }, Ma = async a => {
   a = `${a}.js`;
   let b = await N(a);
@@ -389,17 +389,27 @@ const Va = a => /^[./]/.test(a), Wa = async(a, b, c, d, e = null) => {
       return m;
     }
     m = await m;
-    r = (await Ya(n, b, {nodeModules:c, shallow:d, soft:e, fields:f, package:r || w})).map(z => ({...z, from:z.from ? z.from : n, ...!z.packageJson && p ? {hasMain:p} : {}}));
+    r = (await Ya(n, b, {nodeModules:c, shallow:d, soft:e, fields:f, package:r || w})).map(A => ({...A, from:A.from ? A.from : n, ...!A.packageJson && p ? {hasMain:p} : {}}));
     return [...m, ...r];
   }, g);
 }, Xa = a => Na(/(?:^|[^\w\d_])require\(\s*(['"])(.+?)\1\s*\)/gm, a).map(b => b.from);
 const S = async(a, b = {}) => {
   const c = B();
-  ({path:a} = await Q(a));
+  a = Array.isArray(a) ? a : [a];
+  a = await Promise.all(a.map(async h => {
+    ({path:h} = await Q(h));
+    return h;
+  }));
   const {nodeModules:d = !0, shallow:e = !1, soft:f = !1, fields:g = []} = b;
   let k;
   try {
-    k = await Ya(a, {}, {nodeModules:d, shallow:e, soft:f, fields:g});
+    const h = {};
+    k = await a.reduce(async(l, m) => {
+      l = await l;
+      m = await Ya(m, h, {nodeModules:d, shallow:e, soft:f, fields:g});
+      l.push(...m);
+      return l;
+    }, []);
   } catch (h) {
     throw c(h);
   }
@@ -424,7 +434,7 @@ const S = async(a, b = {}) => {
     m && h ? d.push(m) : m && e.push(m);
     l && g.push(l);
   });
-  return {s:c, A:b, h:d, m:e, l:f, i:g};
+  return {commonJsPackageJsons:c, packageJsons:b, commonJs:d, js:e, internals:f, deps:g};
 };
 const $a = (a, b) => {
   a = " ".repeat(Math.max(a - b.length, 0));
@@ -461,7 +471,7 @@ function bb(a) {
   return `${a}\n--js ${b}`.trim();
 }, eb = async(a, {sourceMap:b}) => {
   const c = [await D(a)];
-  b && (b = H(a), c.push("//" + `# sourceMappingURL=${b}.map`));
+  b && (b = G(a), c.push("//" + `# sourceMappingURL=${b}.map`));
   await E(a, c.join("\n"));
 }, gb = async(a, b = "", c = !1) => {
   if (!b.startsWith("'use strict'") || c) {
@@ -508,47 +518,55 @@ ${a.map(b => {
     c + e.length > b ? (d = d + " \\\n" + e, c = e.length) : (d = d + " " + e, c += e.length + 1);
     return d;
   }, "java");
-}, lb = a => {
+}, lb = async(a, b) => {
+  await Promise.all(a.map(async c => {
+    if (!await N(c)) {
+      throw Error(`Externs file ${c}${b ? ` specified in the "externs" field of package ${b}` : ""} doesn't exist.`);
+    }
+  }));
+}, mb = async a => {
   const b = [];
-  return {v:a.reduce((c, {packageJson:d, externs:e = []}) => {
-    if (!d) {
+  return {files:await a.reduce(async(c, {name:d, packageJson:e, externs:f = []}) => {
+    c = await c;
+    if (!e) {
       return c;
     }
-    const f = J(d);
-    e = Array.isArray(e) ? e : [e];
-    e = e.filter(g => La.includes(g) ? (b.push(g), !1) : !0);
-    d = e.map(g => K(f, g));
-    return [...c, ...d];
-  }, []), ba:b};
-}, U = a => a.reduce((b, c) => [...b, "--externs", c], []), mb = (a, b, c) => c.indexOf(a) == b, nb = (a, b) => a.map(c => c.startsWith(b) ? M(b, c) : c), ob = (a, b, c, d, e, f) => [...a, ...b, ...c && !d ? ["--source_map_include_content"] : [], ...1 < e.length ? ["--module_resolution", "NODE"] : [], ...f ? ["--process_common_js_modules"] : []];
-const pb = require("@depack/nodejs"), qb = (a, b) => {
+    const g = J(e);
+    f = Array.isArray(f) ? f : [f];
+    f = f.filter(k => La.includes(k) ? (b.push(k), !1) : !0);
+    e = f.map(k => K(g, k));
+    await lb(e, d);
+    return [...c, ...e];
+  }, []), V:b};
+}, U = a => a.reduce((b, c) => [...b, "--externs", c], []), nb = (a, b, c) => c.indexOf(a) == b, ob = (a, b) => a.map(c => c.startsWith(b) ? M(b, c) : c), pb = (a, b, c, d, e, f) => [...a, ...b, ...c && !d ? ["--source_map_include_content"] : [], ...1 < e.length ? ["--module_resolution", "NODE"] : [], ...f ? ["--process_common_js_modules"] : []];
+const qb = require("@depack/nodejs"), rb = (a, b) => {
   b = b.split("\n\n").map(c => /^.+?:\d+:(?:\s*Originally at:\s*.+?)? WARNING -/.test(c) ? O(c, "grey") : O(c, "red")).join("\n\n");
   return `Exit code ${a}\n${b}`;
-}, [rb] = process.version.split(".", 1), tb = async({l:a, fa:b = "node_modules", force:c = !0}) => {
-  const d = pb(rb);
+}, [sb] = process.version.split(".", 1), ub = async({internals:a, Z:b = "node_modules", force:c = !0}) => {
+  const d = qb(sb);
   return (await Promise.all(a.map(async e => {
     const f = K(b, e), g = K(f, "package.json");
     var k = K(f, "index.js");
     const h = {packageJson:g, index:k};
     if (await N(g) && !c) {
-      if ((k = await sb(g)) && k == rb) {
+      if ((k = await tb(g)) && k == sb) {
         return h;
       }
       throw Error(`Could not prepare core module ${e}: ${f} exists.`);
     }
     await za(g);
-    await E(g, JSON.stringify({name:e, module:"index.js", depack:rb}));
+    await E(g, JSON.stringify({name:e, module:"index.js", depack:sb}));
     e = await D(K(d, `${e}.js`));
     await E(k, e);
     return h;
   }))).reduce((e, {packageJson:f, index:g}) => [...e, f, g], []);
-}, sb = async a => {
+}, tb = async a => {
   try {
     const b = await D(a), {depack:c} = JSON.parse(b);
     return c;
   } catch (b) {
   }
-}, ub = async(a, b) => {
+}, vb = async(a, b) => {
   a = [...a, ...b];
   await Promise.all(a.map(async c => {
     var d = J(c), e = await D(c);
@@ -566,14 +584,14 @@ const pb = require("@depack/nodejs"), qb = (a, b) => {
     }
     let l, m;
     try {
-      ({aa:l, path:m} = await Q(h, c));
+      ({U:l, path:m} = await Q(h, c));
     } catch (n) {
       throw Error(`The ${k} for dependency ${c} does not exist.`);
     }
     l ? (d = K(h, "index.js"), e[k] = d, console.warn("Updating %s to point to a file.", c), await E(c, JSON.stringify(e, null, 2))) : K(d, e[k]) != m && (d = M(d, m), e[k] = d, console.warn("Updating %s to point to the file with extension.", c), await E(c, JSON.stringify(e, null, 2)));
   }));
 };
-async function vb(a, b) {
+async function wb(a, b) {
   const {interval:c = 250, writable:d = process.stdout} = {writable:process.stderr};
   b = "function" == typeof b ? b() : b;
   const e = d.write.bind(d);
@@ -595,8 +613,8 @@ async function vb(a, b) {
     clearInterval(f), e(`\r${" ".repeat(a.length + 3)}\r`);
   }
 }
-;const {spawn:wb} = child_process;
-const xb = async a => {
+;const {spawn:xb} = child_process;
+const yb = async a => {
   const [b, c, d] = await Promise.all([new Promise((e, f) => {
     a.on("error", f).on("exit", g => {
       e(g);
@@ -604,20 +622,20 @@ const xb = async a => {
   }), a.stdout ? C(a.stdout) : void 0, a.stderr ? C(a.stderr) : void 0]);
   return {code:b, stdout:c, stderr:d};
 };
-function yb(a) {
-  a = wb("java", a, void 0);
-  const b = xb(a);
+function zb(a) {
+  a = xb("java", a, void 0);
+  const b = yb(a);
   a.promise = b;
   a.spawnCommand = a.spawnargs.join(" ");
   return a;
 }
 ;const V = async(a, b = {}) => {
-  const {debug:c, compilerVersion:d, output:e, noSourceMap:f, ca:g} = b;
-  let {promise:k, stderr:h} = yb(a);
+  const {debug:c, compilerVersion:d, output:e, noSourceMap:f, W:g} = b;
+  let {promise:k, stderr:h} = zb(a);
   c && h.pipe(ca(c));
-  const {stdout:l, stderr:m, code:n} = await vb(`Running Google Closure Compiler${d ? " " + O(d, "grey") : ""}`, k);
+  const {stdout:l, stderr:m, code:n} = await wb(`Running Google Closure Compiler${d ? " " + O(d, "grey") : ""}`, k);
   if (n) {
-    throw Error(qb(n, m));
+    throw Error(rb(n, m));
   }
   f || (g ? await Promise.all(g.map(async p => {
     await eb(p, {sourceMap:!0});
@@ -625,20 +643,20 @@ function yb(a) {
   m && !c ? console.warn(O(m, "grey")) : c && console.log("Sources after each pass saved to %s", c);
   return l;
 };
-const zb = require("@depack/externs"), {dependencies:Ab} = zb, Cb = (a, b, c) => {
+const Ab = require("@depack/externs"), {dependencies:Bb} = Ab, Db = (a, b, c) => {
   a = cb([...a, ...b]);
   a = T(a, "js_output_file", "red");
   a = T(a, "externs", "grey");
   a = T(a, "compilation_level", "green", !0);
   console.error(a);
-  const {h:d, l:e, m:f, i:g} = c;
-  c = f.filter(Bb);
-  a = d.filter(Bb);
+  const {commonJs:d, internals:e, js:f, deps:g} = c;
+  c = f.filter(Cb);
+  a = d.filter(Cb);
   g.length && console.error("%s: %s", O("Dependencies", "yellow"), g.filter((k, h, l) => l.indexOf(k) == h).join(" "));
   c.length && console.error("%s: %s", O("Modules", "yellow"), c.join(" "));
   a.length && console.error("%s: %s", O("CommonJS", "yellow"), a.join(" "));
   e.length && console.error("%s: %s", O("Built-ins", "yellow"), e.join(", "));
-}, Eb = a => {
+}, Fb = a => {
   const b = a.map(({hasMain:c, name:d, from:e}) => {
     if (c && d && (c = e.filter(f => {
       const g = a.find(({entry:k}) => k === f);
@@ -646,54 +664,50 @@ const zb = require("@depack/externs"), {dependencies:Ab} = zb, Cb = (a, b, c) =>
         return !0;
       }
     }), c.length)) {
-      return {name:d, X:c};
+      return {name:d, P:c};
     }
   }).filter(Boolean);
-  b.length && (console.error(O(Db(), "red")), console.error("The following commonJS packages referenced in ES6 modules don't support named exports:"), b.forEach(({name:c, X:d}) => {
+  b.length && (console.error(O(Eb(), "red")), console.error("The following commonJS packages referenced in ES6 modules don't support named exports:"), b.forEach(({name:c, P:d}) => {
     console.error(" %s from %s", O(c, "red"), O(d.join(" "), "grey"));
   }));
-}, Db = () => {
+}, Eb = () => {
   let a = "CommonJS don't have named exports, make sure to use them like\nimport myModule from 'my-module' /* CommonJS Compat */\nmyModule.default.method('hello world') // yes Node.JS, wat r u doing\nmyModule.default('must explicitly call default')";
   const b = a.split("\n").reduce((c, {length:d}) => d > c ? d : c, 0);
   process.stderr.isTTY && b + 4 < process.stderr.columns && (a = bb(a));
   return a;
-}, Bb = a => !a.startsWith("node_modules"), Fb = (a, b, c) => c.indexOf(a) == b, Gb = async(a, b = []) => {
-  const c = zb();
-  a = [...[...a, ...b].filter(Fb).reduce((d, e) => {
-    const f = Ab[e] || [];
+}, Cb = a => !a.startsWith("node_modules"), Gb = (a, b, c) => c.indexOf(a) == b, Hb = async(a, b = []) => {
+  const c = Ab();
+  a = [...[...a, ...b].filter(Gb).reduce((d, e) => {
+    const f = Bb[e] || [];
     return [...d, e, ...f];
-  }, []).filter(Fb), "global", "global/buffer", "nodejs"].map(d => {
+  }, []).filter(Gb), "global", "global/buffer", "nodejs"].map(d => {
     ["module", "process", "console", "crypto"].includes(d) && (d = `_${d}`);
     return K(c, `${d}.js`);
   });
-  await Promise.all(a.map(async d => {
-    if (!await N(d)) {
-      throw Error(`Externs ${d} don't exist.`);
-    }
-  }));
+  await lb(a);
   return U(a);
 };
-const {Script:Hb} = vm;
-const Ib = (a, b) => {
+const {Script:Ib} = vm;
+const Jb = (a, b) => {
   const [c, , d] = a.split("\n");
   a = parseInt(c.replace(/.+?(\d+)$/, (f, g) => g)) - 1;
   const e = d.indexOf("^");
   ({length:b} = b.split("\n").slice(0, a).join("\n"));
   return b + e + (a ? 1 : 0);
 };
-const Jb = a => {
+const Kb = a => {
   try {
-    new Hb(a);
+    new Ib(a);
   } catch (b) {
     const {message:c, stack:d} = b;
     if ("Unexpected token <" != c) {
       throw b;
     }
-    return Ib(d, a);
+    return Jb(d, a);
   }
   return null;
 };
-function Kb(a) {
+function Lb(a) {
   if ("object" != typeof a) {
     return !1;
   }
@@ -702,7 +716,7 @@ function Kb(a) {
   const d = -1 != ["string", "function"].indexOf(typeof c);
   return a && d;
 }
-const Lb = (a, b) => {
+const Mb = (a, b) => {
   if (!(b instanceof Error)) {
     throw b;
   }
@@ -718,7 +732,7 @@ const Lb = (a, b) => {
 };
 function W(a, b) {
   function c() {
-    return b.filter(Kb).reduce((d, {re:e, replacement:f}) => {
+    return b.filter(Lb).reduce((d, {re:e, replacement:f}) => {
       if (this.b) {
         return d;
       }
@@ -732,7 +746,7 @@ function W(a, b) {
           try {
             return this.b ? k : f.call(this, k, ...h);
           } catch (l) {
-            Lb(g, l);
+            Mb(g, l);
           }
         });
       }
@@ -743,10 +757,10 @@ function W(a, b) {
   };
   return c.call(c);
 }
-;const Mb = a => new RegExp(`%%_RESTREAM_${a.toUpperCase()}_REPLACEMENT_(\\d+)_%%`, "g"), Nb = (a, b) => `%%_RESTREAM_${a.toUpperCase()}_REPLACEMENT_${b}_%%`, Ob = (a, b) => Object.keys(a).reduce((c, d) => {
+;const Nb = a => new RegExp(`%%_RESTREAM_${a.toUpperCase()}_REPLACEMENT_(\\d+)_%%`, "g"), Ob = (a, b) => `%%_RESTREAM_${a.toUpperCase()}_REPLACEMENT_${b}_%%`, Pb = (a, b) => Object.keys(a).reduce((c, d) => {
   {
     var e = a[d];
-    const {getReplacement:f = Nb, getRegex:g = Mb} = b || {}, k = g(d);
+    const {getReplacement:f = Ob, getRegex:g = Nb} = b || {}, k = g(d);
     e = {name:d, re:e, regExp:k, getReplacement:f, map:{}, lastIndex:0};
   }
   return {...c, [d]:e};
@@ -767,20 +781,20 @@ function W(a, b) {
     return d(e, g);
   }};
 };
-async function Pb(a, b) {
-  return Qb(a, b);
+async function Qb(a, b) {
+  return Rb(a, b);
 }
-class Rb extends ja {
+class Sb extends ja {
   constructor(a, b) {
     super(b);
-    this.a = (Array.isArray(a) ? a : [a]).filter(Kb);
+    this.a = (Array.isArray(a) ? a : [a]).filter(Lb);
     this.b = !1;
-    this.Y = b;
+    this.R = b;
   }
   async replace(a, b) {
-    const c = new Rb(this.a, this.Y);
+    const c = new Sb(this.a, this.R);
     b && Object.assign(c, b);
-    a = await Pb(c, a);
+    a = await Qb(c, a);
     c.b && (this.b = !0);
     b && Object.keys(b).forEach(d => {
       b[d] = c[d];
@@ -808,7 +822,7 @@ class Rb extends ja {
             l instanceof Promise && e.push(l);
             return l;
           } catch (l) {
-            Lb(f, l);
+            Mb(f, l);
           }
         });
         if (e.length) {
@@ -816,7 +830,7 @@ class Rb extends ja {
             const k = await Promise.all(e);
             b = b.replace(c, () => k.shift());
           } catch (k) {
-            Lb(f, k);
+            Mb(f, k);
           }
         } else {
           b = g;
@@ -835,14 +849,14 @@ class Rb extends ja {
     }
   }
 }
-async function Qb(a, b) {
+async function Rb(a, b) {
   b instanceof ia ? b.pipe(a) : a.end(b);
   return await C(a);
 }
-;const Sb = a => {
+;const Tb = a => {
   [, a] = /<\s*(.+?)(?:\s+[\s\S]+)?\s*\/?\s*>/.exec(a) || [];
   return a;
-}, Ub = a => {
+}, Vb = a => {
   let b = 0;
   const c = [];
   let d;
@@ -867,69 +881,69 @@ async function Qb(a, b) {
       throw Error("Could not detect prop name");
     }
     p ? e[p] = l : f.push(l);
-    g[p] = {before:n, G:q, D:r};
+    g[p] = {before:n, v:q, s:r};
     l = h || "";
     l = l.slice(0, l.length - (p || "").length - 1);
-    const {B:w, g:z} = Tb(l);
+    const {m:w, g:A} = Ub(l);
     Object.assign(e, w);
-    Object.assign(g, z);
+    Object.assign(g, A);
     return m + 1;
   }, 0);
   if (c.length) {
     k = a.slice(k);
-    const {B:h, g:l} = Tb(k);
+    const {m:h, g:l} = Ub(k);
     Object.assign(e, h);
     Object.assign(g, l);
   } else {
-    const {B:h, g:l} = Tb(a);
+    const {m:h, g:l} = Ub(a);
     Object.assign(e, h);
     Object.assign(g, l);
   }
-  return {w:e, u:f, g};
-}, Tb = a => {
+  return {l:e, j:f, g};
+}, Ub = a => {
   const b = [], c = {};
   a.replace(/(\s*)(\S+)(\s*)=(\s*)(["'])([\s\S]+?)\5/g, (d, e, f, g, k, h, l, m) => {
-    c[f] = {before:e, G:g, D:k};
-    b.push({j:m, name:f, L:`${h}${l}${h}`});
+    c[f] = {before:e, v:g, s:k};
+    b.push({h:m, name:f, D:`${h}${l}${h}`});
     return "%".repeat(d.length);
   }).replace(/(\s*)([^\s%]+)/g, (d, e, f, g) => {
     c[f] = {before:e};
-    b.push({j:g, name:f, L:"true"});
+    b.push({h:g, name:f, D:"true"});
   });
-  return {B:[...b.reduce((d, {j:e, name:f, L:g}) => {
+  return {m:[...b.reduce((d, {h:e, name:f, D:g}) => {
     d[e] = [f, g];
     return d;
   }, [])].filter(Boolean).reduce((d, [e, f]) => {
     d[e] = f;
     return d;
   }, {}), g:c};
-}, Vb = (a, b = [], c = !1, d = {}, e = "") => {
+}, Wb = (a, b = [], c = !1, d = {}, e = "") => {
   const f = Object.keys(a), {length:g} = f;
   return g || b.length ? `{${f.reduce((k, h) => {
-    const l = a[h], m = c || -1 != h.indexOf("-") ? `'${h}'` : h, {before:n = "", G:p = "", D:q = ""} = d[h] || {};
+    const l = a[h], m = c || -1 != h.indexOf("-") ? `'${h}'` : h, {before:n = "", v:p = "", s:q = ""} = d[h] || {};
     return [...k, `${n}${m}${p}:${q}${l}`];
   }, b).join(",")}${e}}` : "{}";
-}, Wb = (a = "") => {
+}, Xb = (a = "") => {
   [a] = a;
   if (!a) {
     throw Error("No tag name is given");
   }
   return a.toUpperCase() == a;
-}, Xb = (a, b = {}, c = [], d = [], e = !1, f = null, g = {}, k = "") => {
-  const h = Wb(a), l = h ? a : `'${a}'`;
+}, Yb = (a, b = {}, c = [], d = [], e = !1, f = null, g = {}, k = "") => {
+  const h = Xb(a), l = h ? a : `'${a}'`;
   if (!Object.keys(b).length && !c.length && !d.length) {
     return `h(${l})`;
   }
   const m = h && "dom" == e ? !1 : e;
   h || !d.length || e && "dom" != e || f && f(`JSX: destructuring ${d.join(" ")} is used without quoted props on HTML ${a}.`);
-  a = Vb(b, d, m, g, k);
+  a = Wb(b, d, m, g, k);
   b = c.reduce((n, p, q) => {
     q = c[q - 1];
     return `${n}${q && /\S/.test(q) ? "," : ""}${p}`;
   }, "");
   return `h(${l},${a}${b ? `,${b}` : ""})`;
 };
-const Yb = (a, b = []) => {
+const Zb = (a, b = []) => {
   let c = 0, d;
   a = W(a, [...b, {re:/[<>]/g, replacement(e, f) {
     if (d) {
@@ -943,13 +957,13 @@ const Yb = (a, b = []) => {
   if (c) {
     throw Error(1);
   }
-  return {da:a, J:d};
-}, $b = a => {
-  const b = Sb(a);
+  return {X:a, B:d};
+}, ac = a => {
+  const b = Tb(a);
   let c;
-  const {N:d} = Ob({N:/=>/g});
+  const {G:d} = Pb({G:/=>/g});
   try {
-    ({da:h, J:c} = Yb(a, [Z(d)]));
+    ({X:h, B:c} = Zb(a, [Z(d)]));
   } catch (l) {
     if (1 === l) {
       throw Error(`Could not find the matching closing > for ${b}.`);
@@ -958,7 +972,7 @@ const Yb = (a, b = []) => {
   const e = h.slice(0, c + 1);
   var f = e.replace(/<\s*[^\s/>]+/, "");
   if (/\/\s*>$/.test(f)) {
-    return a = f.replace(/\/\s*>$/, ""), f = "", new Zb({f:e.replace(d.regExp, "=>"), c:a.replace(d.regExp, "=>"), content:"", tagName:b});
+    return a = f.replace(/\/\s*>$/, ""), f = "", new $b({f:e.replace(d.regExp, "=>"), c:a.replace(d.regExp, "=>"), content:"", tagName:b});
   }
   a = f.replace(/>$/, "");
   f = c + 1;
@@ -972,7 +986,7 @@ const Yb = (a, b = []) => {
     const q = !m;
     if (q) {
       p = p.slice(n);
-      const {J:r} = Yb(p.replace(/^[\s\S]/, " "));
+      const {B:r} = Zb(p.replace(/^[\s\S]/, " "));
       p = p.slice(0, r + 1);
       if (/\/\s*>$/.test(p)) {
         return l;
@@ -987,9 +1001,9 @@ const Yb = (a, b = []) => {
   }
   f = h.slice(f, c);
   var h = h.slice(0, k).replace(d.regExp, "=>");
-  return new Zb({f:h, c:a.replace(d.regExp, "=>"), content:f.replace(d.regExp, "=>"), tagName:b});
+  return new $b({f:h, c:a.replace(d.regExp, "=>"), content:f.replace(d.regExp, "=>"), tagName:b});
 };
-class Zb {
+class $b {
   constructor(a) {
     this.f = a.f;
     this.c = a.c;
@@ -997,7 +1011,7 @@ class Zb {
     this.tagName = a.tagName;
   }
 }
-;const ac = a => {
+;const bc = a => {
   let b = "", c = "";
   a = a.replace(/^(\n\s*)([\s\S]+)?/, (d, e, f = "") => {
     b = e;
@@ -1007,110 +1021,110 @@ class Zb {
     return e;
   });
   return `${b}${a ? `\`${a}\`` : ""}${c}`;
-}, cc = a => {
+}, dc = a => {
   const b = [];
   let c = {}, d = 0, e = 0;
   W(a, [{re:/[<{}]/g, replacement(f, g) {
     if (!(g < e)) {
       if (/[{}]/.test(f)) {
-        d += "{" == f ? 1 : -1, 1 == d && void 0 == c.from ? c.from = g : 0 == d && (c.C = g + 1, c.T = a.slice(c.from + 1, g), b.push(c), c = {});
+        d += "{" == f ? 1 : -1, 1 == d && void 0 == c.from ? c.from = g : 0 == d && (c.o = g + 1, c.L = a.slice(c.from + 1, g), b.push(c), c = {});
       } else {
         if (d) {
           return f;
         }
-        f = $b(a.slice(g));
+        f = ac(a.slice(g));
         e = g + f.f.length;
-        c.U = f;
-        c.C = e;
+        c.M = f;
+        c.o = e;
         c.from = g;
         b.push(c);
         c = {};
       }
     }
   }}, {}]);
-  return b.length ? bc(a, b) : [ac(a)];
-}, bc = (a, b) => {
+  return b.length ? cc(a, b) : [bc(a)];
+}, cc = (a, b) => {
   let c = 0;
-  b = b.reduce((d, {from:e, C:f, T:g, U:k}) => {
-    (e = a.slice(c, e)) && d.push(ac(e));
+  b = b.reduce((d, {from:e, o:f, L:g, M:k}) => {
+    (e = a.slice(c, e)) && d.push(bc(e));
     c = f;
     g ? d.push(g) : k && d.push(k);
     return d;
   }, []);
   if (c < a.length) {
     const d = a.slice(c, a.length);
-    d && b.push(ac(d));
+    d && b.push(bc(d));
   }
   return b;
 };
-const ec = (a, b = {}) => {
+const fc = (a, b = {}) => {
   const {quoteProps:c, warn:d} = b;
-  var e = Jb(a);
+  var e = Kb(a);
   if (null === e) {
     return a;
   }
   var f = a.slice(e);
-  const {c:g = "", content:k, tagName:h, f:{length:l}} = $b(f);
-  f = dc(k, c, d);
-  const {w:m, u:n, g:p} = Ub(g.replace(/^ */, ""));
-  var q = Xb(h, m, f, n, c, d, p, /\s*$/.exec(g) || [""]);
+  const {c:g = "", content:k, tagName:h, f:{length:l}} = ac(f);
+  f = ec(k, c, d);
+  const {l:m, j:n, g:p} = Vb(g.replace(/^ */, ""));
+  var q = Yb(h, m, f, n, c, d, p, /\s*$/.exec(g) || [""]);
   f = a.slice(0, e);
   a = a.slice(e + l);
   e = l - q.length;
   0 < e && (q = `${" ".repeat(e)}${q}`);
   f = `${f}${q}${a}`;
-  return ec(f, b);
-}, dc = (a, b = !1, c = null) => a ? cc(a).reduce((d, e) => {
-  if (e instanceof Zb) {
-    const {c:k = "", content:h, tagName:l} = e, {w:m, u:n} = Ub(k);
-    e = dc(h, b, c);
-    e = Xb(l, m, e, n, b, c);
+  return fc(f, b);
+}, ec = (a, b = !1, c = null) => a ? dc(a).reduce((d, e) => {
+  if (e instanceof $b) {
+    const {c:k = "", content:h, tagName:l} = e, {l:m, j:n} = Vb(k);
+    e = ec(h, b, c);
+    e = Yb(l, m, e, n, b, c);
     return [...d, e];
   }
-  const f = Jb(e);
+  const f = Kb(e);
   if (f) {
     var g = e.slice(f);
-    const {f:{length:k}, c:h = "", content:l, tagName:m} = $b(g), {w:n, u:p} = Ub(h);
-    g = dc(l, b, c);
-    g = Xb(m, n, g, p, b, c);
+    const {f:{length:k}, c:h = "", content:l, tagName:m} = ac(g), {l:n, j:p} = Vb(h);
+    g = ec(l, b, c);
+    g = Yb(m, n, g, p, b, c);
     const q = e.slice(0, f);
     e = e.slice(f + k);
     return [...d, `${q}${g}${e}`];
   }
   return [...d, e];
 }, []) : [];
-const fc = (a, b = {}) => {
-  const {e:c, P:d, R:e, j:f, Z:g, $:k} = Ob({P:/^ *export\s+default\s+{[\s\S]+?}/mg, e:/^ *export\s+(?:default\s+)?/mg, R:/^ *export\s+{[^}]+}\s+from\s+(['"])(?:.+?)\1/mg, j:/^ *import(\s+([^\s,]+)\s*,?)?(\s*{(?:[^}]+)})?\s+from\s+['"].+['"]/gm, Z:/^ *import\s+(?:(.+?)\s*,\s*)?\*\s+as\s+.+?\s+from\s+['"].+['"]/gm, $:/^ *import\s+['"].+['"]/gm}, {getReplacement(h, l) {
+const gc = (a, b = {}) => {
+  const {e:c, I:d, J:e, h:f, S:g, T:k} = Pb({I:/^ *export\s+default\s+{[\s\S]+?}/mg, e:/^ *export\s+(?:default\s+)?/mg, J:/^ *export\s+{[^}]+}\s+from\s+(['"])(?:.+?)\1/mg, h:/^ *import(\s+([^\s,]+)\s*,?)?(\s*{(?:[^}]+)})?\s+from\s+['"].+['"]/gm, S:/^ *import\s+(?:(.+?)\s*,\s*)?\*\s+as\s+.+?\s+from\s+['"].+['"]/gm, T:/^ *import\s+['"].+['"]/gm}, {getReplacement(h, l) {
     return `/*%%_RESTREAM_${h.toUpperCase()}_REPLACEMENT_${l}_%%*/`;
   }, getRegex(h) {
     return new RegExp(`/\\*%%_RESTREAM_${h.toUpperCase()}_REPLACEMENT_(\\d+)_%%\\*/`, "g");
   }});
   a = W(a, [Z(e), Z(d), Z(c), Z(f), Z(g), Z(k)]);
-  b = ec(a, b);
+  b = fc(a, b);
   return W(b, [Y(e), Y(d), Y(c), Y(f), Y(g), Y(k)]);
 };
-class gc extends Rb {
+class hc extends Sb {
   constructor(a, b) {
     super([]);
     const c = this.replacement.bind(this);
     this.a = [{re:/^( *import(?:\s+[^\s,]+\s*,?)?(?:\s*{(?:[^}]+)})?\s+from\s+)['"](.+)['"]/gm, replacement:c}, {re:/^( *import\s+['"](.+)['"])/gm, replacement:c}, {re:/^( *export\s+{[^}]+?}\s+from\s+)['"](.+?)['"]/gm, replacement:c}];
-    this.V = [];
-    this.F = [];
-    this.O = [];
+    this.N = [];
+    this.u = [];
+    this.H = [];
     this.path = a;
-    this.C = b;
+    this.o = b;
     this.preactExtern = !1;
   }
   get nodeModules() {
-    return this.V;
+    return this.N;
   }
-  get i() {
-    return this.O;
+  get deps() {
+    return this.H;
   }
   async replacement(a, b, c) {
     var d = J(this.path);
     if (c.endsWith(".css")) {
-      return this.F.push(c), a;
+      return this.u.push(c), a;
     }
     if (/^[./]/.test(c)) {
       var {path:e} = await Q(c, this.path);
@@ -1133,7 +1147,7 @@ class gc extends Rb {
         }
         f && (c = K("node_modules", f), c = M(d, c));
       }
-      this.i.push(c);
+      this.deps.push(c);
       d = c.startsWith(".") ? "" : "./";
       return a == b ? b.replace(/(['"]).+\1/, `$1${d}${c.replace(/(\/index)?\.js$/, "")}$1`) : `${b}'${d}${c.replace(/(\/index)?\.js$/, "")}'`;
     }
@@ -1141,15 +1155,15 @@ class gc extends Rb {
     return "preact" == c && this.preactExtern ? ({entry:a} = await R(d, "@externs/preact"), this.nodeModules.push(a), `${b}'@externs/preact'`) : a;
   }
 }
-;const ic = async(a, b, c) => {
-  const {I:d, H:e} = c, {tempDir:f, preact:g, preactExtern:k} = b;
+;const jc = async(a, b, c) => {
+  const {A:d, w:e} = c, {tempDir:f, preact:g, preactExtern:k} = b;
   var h = await D(a), l = a.endsWith(".jsx");
-  const m = M("", J(a)), n = K(f, m), p = new gc(a, n);
+  const m = M("", J(a)), n = K(f, m), p = new hc(a, n);
   p.preactExtern = k;
   p.end((g || k) && l ? `import { h } from '${k ? "@externs/preact" : "preact"}'
 ${h}` : h);
   h = await C(p);
-  l = l ? await hc(h, a) : h;
+  l = l ? await ic(h, a) : h;
   if (a.startsWith("..")) {
     for (h = a; "." != h && !q;) {
       h = J(h);
@@ -1164,7 +1178,7 @@ ${h}` : h);
   a = K(f, a);
   await za(a);
   await E(a, l);
-  a = p.i.map(r => K(m, r)).filter(r => !(r in e));
+  a = p.deps.map(r => K(m, r)).filter(r => !(r in e));
   q = p.nodeModules.filter(r => !(r in d));
   q.forEach(r => {
     d[r] = 1;
@@ -1172,14 +1186,11 @@ ${h}` : h);
   a.forEach(r => {
     e[r] = 1;
   });
-  await q.reduce(async(r, w) => {
-    await r;
-    (await S(w)).forEach(({entry:z, packageJson:G}) => {
-      G && (d[G] = 1);
-      d[z] = 1;
-    });
-  }, {});
-  await p.F.reduce(async(r, w) => {
+  (await S(q)).forEach(({entry:r, packageJson:w}) => {
+    w && (d[w] = 1);
+    d[r] = 1;
+  });
+  await p.u.reduce(async(r, w) => {
     await r;
     r = K(m, w);
     r = `import injectStyle from 'depack/inject-css'
@@ -1190,33 +1201,41 @@ injectStyle(\`${await D(r)}\`)`;
   }, {});
   await a.reduce(async(r, w) => {
     await r;
-    await ic(w, b, c);
+    await jc(w, b, c);
   }, {});
-}, hc = async(a, b) => await fc(a, {quoteProps:"dom", warn(c) {
+}, ic = async(a, b) => await gc(a, {quoteProps:"dom", warn(c) {
   console.warn(O(c, "yellow"));
   console.log(b);
 }});
-const jc = async(a, b = {}) => {
+const kc = async(a, b = {}) => {
   const {tempDir:c = "depack-temp", preact:d, preactExtern:e} = b;
-  b = {H:{[M("", a)]:1}, I:{}};
-  await ic(a, {tempDir:c, preact:d, preactExtern:e}, b);
-  return [...Object.keys(b.H).map(f => K(c, f)), ...Object.keys(b.I)];
+  b = {w:{[M("", a)]:1}, A:{}};
+  await jc(a, {tempDir:c, preact:d, preactExtern:e}, b);
+  return [...Object.keys(b.w).map(f => K(c, f)), ...Object.keys(b.A)];
 };
-const lc = async a => Array.isArray(a) ? a.reduce(async(b, c) => (b = await b) ? b : await kc(c), !1) : await kc(a), kc = async a => {
-  const b = await S(a, {nodeModules:!1});
-  return a.endsWith(".jsx") || b.some(({entry:c}) => c.endsWith(".jsx"));
-}, mc = async(a, {tempDir:b, preact:c, preactExtern:d, W:e}) => {
+const mc = async a => {
+  if (Array.isArray(a)) {
+    if (a.some(lc)) {
+      return !0;
+    }
+  } else {
+    if (a.endsWith(".jsx")) {
+      return !0;
+    }
+  }
+  return (await S(a, {nodeModules:!1})).some(({entry:b}) => b.endsWith(".jsx"));
+}, lc = a => a.endsWith(".jsx"), nc = async(a, {tempDir:b, preact:c, preactExtern:d, O:e}) => {
   let f = a;
   if (e) {
-    return await jc(a, {tempDir:b, preact:c, preactExtern:d}), f = K(b, a), {o:f, K:!0};
+    return await kc(a, {tempDir:b, preact:c, preactExtern:d}), f = K(b, a), {i:f, C:!0};
   }
-  if (e = await lc(a)) {
-    await jc(a, {tempDir:b, preact:c, preactExtern:d}), f = K(b, a);
+  if (e = await mc(a)) {
+    await kc(a, {tempDir:b, preact:c, preactExtern:d}), f = K(b, a);
   }
-  return {o:f, K:e};
+  return {i:f, C:e};
 };
-const nc = (a, b) => [...a, "--js", b];
-const oc = process.env.GOOGLE_CLOSURE_COMPILER;
+const oc = (a, b) => [...a, "--js", b];
+const pc = process.env.GOOGLE_CLOSURE_COMPILER;
 /*
 
  @depack/depack: Depack Node.JS API for Closure Compiler execution.
@@ -1242,38 +1261,38 @@ module.exports = {_Compile:async(a, b = {}, c = []) => {
   if (!d) {
     throw Error("Source is not given.");
   }
-  var k = c.reduce((u, A, I, L) => {
-    if ("--externs" != A) {
-      return u;
+  var k = c.reduce((v, z, H, L) => {
+    if ("--externs" != z) {
+      return v;
     }
-    A = L[I + 1];
-    if (!A) {
-      return u;
+    z = L[H + 1];
+    if (!z) {
+      return v;
     }
-    La.includes(A) && (c[I] = "", c[I + 1] = "", u.push(A));
-    return u;
+    La.includes(z) && (c[H] = "", c[H + 1] = "", v.push(z));
+    return v;
   }, []);
-  const h = [...k.length ? c.filter(u => u) : c, "--package_json_entry_names", "module,main", "--entry_point", d];
+  const h = [...k.length ? c.filter(v => v) : c, "--package_json_entry_names", "module,main", "--entry_point", d];
   var l = await S(d, {fields:["externs"]});
-  const {v:m, ba:n} = lb(l);
+  const {files:m, V:n} = await mb(l);
   m.length && console.error("%s %s", O("Modules' externs:", "blue"), m.join(" "));
   const p = U(m);
-  Eb(l);
-  const q = Za(l), {h:r, s:w, l:z, m:G, A:P} = q;
-  var v = await tb({l:z});
-  k = await Gb(z, [...k, ...n]);
-  await ub(w, P);
-  var x = [d, ...w, ...P, ...G, ...r, ...v].sort((u, A) => u.startsWith("node_modules") ? -1 : A.startsWith("node_modules") ? 1 : 0);
-  v = ib(z);
+  Fb(l);
+  const q = Za(l), {commonJs:r, commonJsPackageJsons:w, internals:A, js:I, packageJsons:P} = q;
+  var u = await ub({internals:A});
+  k = await Hb(A, [...k, ...n]);
+  await vb(w, P);
+  var x = [d, ...w, ...P, ...I, ...r, ...u].sort((v, z) => v.startsWith("node_modules") ? -1 : z.startsWith("node_modules") ? 1 : 0);
+  u = ib(A);
   l = jb(l);
-  x = [...h, ...k, ...p, ...1 < x.length ? ["--module_resolution", "NODE"] : [], ...r.length ? ["--process_common_js_modules"] : [], ...v ? ["--output_wrapper", v] : [], "--js", ...x];
-  l.length && !r.length && (l = l.filter(({required:u}) => u, !1), l.length && (console.error("You are requiring JSON files. Make sure their relative paths will stay the same to the build."), console.log(l.map(({entry:u, from:A}) => `${O(u, "blue")} from ${A.join(" ")}`).join("\n"))));
-  f ? console.error(cb(x)) : Cb(h, [...k, ...p], q);
+  x = [...h, ...k, ...p, ...1 < x.length ? ["--module_resolution", "NODE"] : [], ...r.length ? ["--process_common_js_modules"] : [], ...u ? ["--output_wrapper", u] : [], "--js", ...x];
+  l.length && !r.length && (l = l.filter(({required:v}) => v, !1), l.length && (console.error("You are requiring JSON files. Make sure their relative paths will stay the same to the build."), console.log(l.map(({entry:v, from:z}) => `${O(v, "blue")} from ${z.join(" ")}`).join("\n"))));
+  f ? console.error(cb(x)) : Db(h, [...k, ...p], q);
   b = await V(x, b);
   if (!a) {
-    return b = fb(b, v, e).trim(), g || console.log(b), b;
+    return b = fb(b, u, e).trim(), g || console.log(b), b;
   }
-  await gb(a, v, e);
+  await gb(a, u, e);
   await F(aa, [a, "755"]);
   return b;
 }, _Bundle:async(a, b = {}, c = []) => {
@@ -1281,21 +1300,21 @@ module.exports = {_Compile:async(a, b = {}, c = []) => {
   if (!d) {
     throw Error("Entry file is not given.");
   }
-  let {o:p, K:q} = await mc(d, {tempDir:e, preact:f, preactExtern:g});
+  let {i:p, C:q} = await nc(d, {tempDir:e, preact:f, preactExtern:g});
   a = await S(p, {fields:["externs"]});
-  ({v:b} = lb(a));
+  ({files:b} = await mb(a));
   b = U(b);
   var r = Za(a);
-  const {h:w, s:z, m:G, A:P} = r;
+  const {commonJs:w, commonJsPackageJsons:A, js:I, packageJsons:P} = r;
   a = jb(a);
   r = !(!w.length && !a.length);
-  a = [p, ...w, ...P, ...G, ...z];
-  c = ob(c, b, h, n, a, r);
-  b = q ? a.map(v => v.startsWith(e) ? M(e, v) : v) : a;
+  a = [p, ...w, ...P, ...I, ...A];
+  c = pb(c, b, h, n, a, r);
+  b = q ? a.map(u => u.startsWith(e) ? M(e, u) : u) : a;
   b = db(c, b);
   console.error(b);
   c = [...c, "--js", ...a];
-  c = await V(c, {debug:m, compilerVersion:l, output:h, noSourceMap:n, ea:() => !1});
+  c = await V(c, {debug:m, compilerVersion:l, output:h, noSourceMap:n, Y:() => !1});
   h || !c || k || console.log(c);
   q && (h && !n && await hb(h, e), await Ha(e));
   return c;
@@ -1307,49 +1326,51 @@ module.exports = {_Compile:async(a, b = {}, c = []) => {
   if (!Array.isArray(d)) {
     throw Error("Expecting an array of source files to generate chunks.");
   }
-  let n = [], p = !1, q = await lc(d), r = [];
+  let n = [], p = !1, q = await mc(d), r = [];
   const w = {};
-  b = await d.reduce(async(v, x) => {
-    v = await v;
-    ({o:x} = await mc(x, {tempDir:e, preact:f, preactExtern:g, W:q}));
-    var u = await S(x, {fields:["externs"]}), {v:A} = lb(u);
-    r = [...r, ...A];
-    A = Za(u);
-    const {h:I, s:L, m:X, A:ta} = A;
-    u = jb(u);
-    p = p || !(!I.length && !u.length);
-    u = [...I, ...ta, ...X, ...L];
-    n = [...n, ...u];
-    v[x] = u;
-    return v;
+  b = await d.reduce(async(u, x) => {
+    u = await u;
+    ({i:x} = await nc(x, {tempDir:e, preact:f, preactExtern:g, O:q}));
+    var v = await S(x, {fields:["externs"]}), {files:z} = await mb(v);
+    r = [...r, ...z];
+    z = Za(v);
+    const {commonJs:H, commonJsPackageJsons:L, js:X, packageJsons:ta} = z;
+    v = jb(v);
+    p = p || !(!H.length && !v.length);
+    v = [...H, ...ta, ...X, ...L];
+    n = [...n, ...v];
+    u[x] = v;
+    return u;
   }, {});
-  const z = n.reduce((v, x) => {
-    v[x] ? v[x]++ : v[x] = 1;
-    return v;
+  const A = n.reduce((u, x) => {
+    u[x] ? u[x]++ : u[x] = 1;
+    return u;
   }, {});
-  a = Object.entries(z).reduce((v, [x, u]) => {
-    1 < u && v.push("--js", x);
-    return v;
+  a = Object.entries(A).reduce((u, [x, v]) => {
+    1 < v && u.push("--js", x);
+    return u;
   }, []);
-  a.length && a.push("--chunk", `common:${a.length / 2}`);
-  const G = [];
-  b = Object.entries(b).reduce((v, [x, u]) => {
-    const A = u.filter(ta => 1 == z[ta]), I = A.reduce(nc, []), L = H(x).replace(/.jsx$/, ".js"), X = [L.replace(".js", ""), A.length + 1];
-    A.length != u.length && (w[L] = ["common"], X.push("common"));
-    v.push(...I, "--js", x, "--chunk", X.join(":"));
+  const I = [];
+  a.length && (a.push("--chunk", `common:${a.length / 2}`), I.push(K(k, "common.js")));
+  b = Object.entries(b).reduce((u, [x, v]) => {
+    const z = v.filter(ta => 1 == A[ta]), H = z.reduce(oc, []), L = G(x).replace(/.jsx$/, ".js"), X = [L.replace(".js", ""), z.length + 1];
+    z.length != v.length && (w[L] = ["common"], X.push("common"));
+    u.push(...H, "--js", x, "--chunk", X.join(":"));
     x = K(k, L);
-    G.push(x);
-    return v;
+    I.push(x);
+    return u;
   }, []);
-  const P = U(r.filter(mb));
-  c = ob(c, P, k, m, n, p);
+  const P = U(r.filter(nb));
+  c = pb(c, P, k, m, n, p);
   a = [...a, ...b];
-  b = db(c, nb(a, e));
+  b = db(c, ob(a, e));
   console.error(b);
   c = [...c, ...a];
-  c = await V(c, {debug:l, compilerVersion:h, output:k, noSourceMap:m, ca:G});
+  c = await V(c, {debug:l, compilerVersion:h, output:k, noSourceMap:m, W:I});
   !k && c && console.log(c);
-  q && await Ha(e);
+  q && (k && !m && await Promise.all(I.map(async u => {
+    await hb(u, e);
+  })), await Ha(e));
   return w;
 }, _run:V, _getOptions:(a = {}) => {
   const {compiler:b = require.resolve("google-closure-compiler-java/compiler.jar"), output:c, level:d, advanced:e, languageIn:f, languageOut:g, sourceMap:k = !0, argv:h = [], prettyPrint:l, noWarnings:m, debug:n, iife:p, chunkOutput:q} = a;
@@ -1367,14 +1388,14 @@ module.exports = {_Compile:async(a, b = {}, c = []) => {
   q && a.push("--chunk_output_path_prefix", K(q, ya));
   return a;
 }, _getOutput:(a, b) => {
-  a = /\.js$/.test(a) ? a : K(a, H(b));
+  a = /\.js$/.test(a) ? a : K(a, G(b));
   return a = a.replace(/jsx$/, "js");
 }, _getCompilerVersion:async() => {
   var a = "target";
-  const b = oc ? "target" : require.resolve("google-closure-compiler-java/package.json");
-  oc || (a = await D(b), {version:a} = JSON.parse(a), [a] = a.split("."));
+  const b = pc ? "target" : require.resolve("google-closure-compiler-java/package.json");
+  pc || (a = await D(b), {version:a} = JSON.parse(a), [a] = a.split("."));
   return a;
-}, _GOOGLE_CLOSURE_COMPILER:oc};
+}, _GOOGLE_CLOSURE_COMPILER:pc};
 
 
 //# sourceMappingURL=depack.js.map