+++ src/lib/compile.js
@@ -3,11 +3,9 @@ import { join } from 'path'
 import makePromise from 'makepromise'
 import { chmod } from 'fs'
 import { builtinModules } from 'module'
-import { exists } from '@wrote/wrote'
 import detect, { sort } from 'static-analysis'
-// import getExternsDir, { dependencies as externsDeps } from '@depack/externs'
 import frame from 'frame-of-mind'
-import { removeStrict, getWrapper, hasJsonFiles, prepareOutput, getShellCommand, replaceWithColor, detectExterns, createExternsArgs } from './'
+import { removeStrict, getWrapper, hasJsonFiles, prepareOutput, getShellCommand, replaceWithColor, detectExterns, createExternsArgs, checkExternsExist } from './'
 import { prepareCoreModules, fixDependencies } from './closure'
 import run from './run'
 
@@ -52,7 +50,7 @@ const Compile = async (options, runOptions = {}, compilerArgs = []) => {
   const detected = await detect(src, {
     fields: ['externs'],
   })
-  const { detectedExterns, nodeJS } = detectExterns(detected)
+  const { files: detectedExterns, nodeJS } = await detectExterns(detected)
   detectedExterns.length && console.error('%s %s', c('Modules\' externs:', 'blue'), detectedExterns.join(' '))
   const detectedExternsArgs = createExternsArgs(detectedExterns)
   warnOfCommonJs(detected)
@@ -62,7 +60,7 @@ const Compile = async (options, runOptions = {}, compilerArgs = []) => {
     commonJs, commonJsPackageJsons, internals, js, packageJsons,
   } = sorted
   const internalDeps = await prepareCoreModules({ internals })
-  const externs = await getExterns(internals, [
+  const externs = await getNodeExterns(internals, [
     ...foundAdditional,
     ...nodeJS,
   ])
@@ -192,11 +190,11 @@ const filterNodeModule = (entry) => {
 const unique = (e, i, a) => a.indexOf(e) == i
 
 /**
- * Returns options to include externs.
+ * Returns compiler arguments to include Node.JS externs.
  * @param {!Array<string>} internals The list of builtin Node.JS modules used.
  * @param {!Array<string>} additional Any extra Node.JS modules to be included
  */
-export const getExterns = async (internals, additional = []) => {
+export const getNodeExterns = async (internals, additional = []) => {
   const externsDir = getExternsDir()
   const allInternals = [...internals, ...additional]
     .filter(unique)
@@ -205,16 +203,13 @@ export const getExterns = async (internals, additional = []) => {
       return [...acc, i, ...deps]
     }, [])
     .filter(unique)
-  const p = [...allInternals, 'global', 'global/buffer', 'nodejs']
+  const internalExterns = [...allInternals, 'global', 'global/buffer', 'nodejs']
     .map(i => {
       if (['module', 'process', 'console', 'crypto'].includes(i)) i = `_${i}`
       return join(externsDir, `${i}.js`)
     })
-  await Promise.all(p.map(async pp => {
-    const exist = await exists(pp)
-    if (!exist) throw new Error(`Externs ${pp} don't exist.`)
-  }))
-  const args = createExternsArgs(p)
+  await checkExternsExist(internalExterns)
+  const args = createExternsArgs(internalExterns)
   return args
 }